+++ mach_eval/analyzers/electromagnetic/bspm/jmag_2d.py
